storybook restructure
